Back out "Add rocprofiler-sdk support 2" (#1162)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/1162

Original commit changeset: 56db042d067c

Original Phabricator Diff: D82773951

Reviewed By: haoyuz

Differential Revision: D85735557

fbshipit-source-id: 3736c3d3df76953b1ba68fa41c33ad31995d4605
