+++ libkineto/CMakeLists.txt
@@ -79,50 +79,6 @@ else()
   set(LIBKINETO_NOXPUPTI ON)
 endif()
 
-# Detect ROCM Version
-if(NOT LIBKINETO_NOROCTRACER)
-  if(NOT ROCM_INCLUDE_DIRS)
-    set(ROCM_INCLUDE_DIRS "${ROCM_SOURCE_DIR}/include")
-  endif()
-
-  find_file(ROCM_VERSION_HEADER_PATH
-    NAMES rocm-core/rocm_version.h
-    NO_DEFAULT_PATH
-    PATHS ${ROCM_INCLUDE_DIRS}
-  )
-
-  if(EXISTS ${ROCM_VERSION_HEADER_PATH})
-    set(ROCM_HEADER_FILE ${ROCM_VERSION_HEADER_PATH})
-  endif()
-
-  # Read the ROCM headerfile into a variable
-  message(STATUS "Reading ROCM version from: ${ROCM_HEADER_FILE}")
-  file(READ "${ROCM_HEADER_FILE}" ROCM_HEADER_CONTENT)
-
-  string(REGEX MATCH "ROCM_VERSION_MAJOR[ ]+[0-9]+" TEMP1 ${ROCM_HEADER_CONTENT})
-  string(REPLACE "ROCM_VERSION_MAJOR" "" TEMP2 ${TEMP1})
-  string(STRIP ${TEMP2} ROCM_VERSION_DEV_MAJOR)
-  string(REGEX MATCH "ROCM_VERSION_MINOR[ ]+[0-9]+" TEMP1 ${ROCM_HEADER_CONTENT})
-  string(REPLACE "ROCM_VERSION_MINOR" "" TEMP2 ${TEMP1})
-  string(STRIP ${TEMP2} ROCM_VERSION_DEV_MINOR)
-  string(REGEX MATCH "ROCM_VERSION_PATCH[ ]+[0-9]+" TEMP1 ${ROCM_HEADER_CONTENT})
-  string(REPLACE "ROCM_VERSION_PATCH" "" TEMP2 ${TEMP1})
-  string(STRIP ${TEMP2} ROCM_VERSION_DEV_PATCH)
-
-  message(STATUS "ROCM major: ${ROCM_VERSION_DEV_MAJOR}")
-  message(STATUS "ROCM minor: ${ROCM_VERSION_DEV_MINOR}")
-  message(STATUS "ROCM patch: ${ROCM_VERSION_DEV_PATCH}")
-
-# Use rocprofiler-sdk for rocm version 6.4 forward
-  if ((${ROCM_VERSION_DEV_MAJOR} GREATER_EQUAL 7) OR (${ROCM_VERSION_DEV_MAJOR} GREATER_EQUAL 6 AND ${ROCM_VERSION_DEV_MINOR} GREATER_EQUAL 4))
-    set(USE_ROCPROFILER_SDK ON)
-  endif()
-  if (${USE_ROCPROFILER_SDK})
-    message(STATUS "Building with: rocprofiler-sdk")
-  else()
-    message(STATUS "Building with: libroctracer")
-  endif()
-endif()
 if(NOT DEFINED LIBKINETO_NOAIUPTI)
   message(INFO " LIBKINETO_NOAIUPTI NOT DEFINED adding subdirectory(src/plugin/aiupti)")
   add_subdirectory(src/plugin/aiupti)
@@ -134,14 +90,8 @@ if(LIBKINETO_NOCUPTI AND LIBKINETO_NOROCTRACER AND LIBKINETO_NOXPUPTI AND LIBKIN
   message(STATUS " CUPTI unavailable or disabled - not building GPU profilers")
 else()
   if(NOT LIBKINETO_NOROCTRACER)
-    if (${USE_ROCPROFILER_SDK})
-      get_filelist("get_libkineto_rocprofiler_srcs(with_api=False)" LIBKINETO_roc_SRCS)
-      message(STATUS " Building with rocprofiler-sdk")
-    else()
-      get_filelist("get_libkineto_roctracer_srcs(with_api=False)" LIBKINETO_roc_SRCS)
-      add_compile_options(-DROCTRACER_FALLBACK)
-      message(STATUS " Building with roctracer")
-    endif()
+    get_filelist("get_libkineto_roctracer_srcs(with_api=False)" LIBKINETO_roc_SRCS)
+    message(STATUS " Building with roctracer")
   elseif(NOT LIBKINETO_NOCUPTI)
     get_filelist("get_libkineto_cupti_srcs(with_api=False)" LIBKINETO_cuda_SRCS)
   endif()
@@ -278,14 +228,9 @@ target_include_directories(kineto PUBLIC
       $<BUILD_INTERFACE:${LIBKINETO_SOURCE_DIR}>)
 
 if(NOT LIBKINETO_NOROCTRACER)
-if (${USE_ROCPROFILER_SDK})
-  find_library(ROCPROF_LIBRARY NAMES librocprofiler-sdk.so HINTS
+  find_library(ROCTRACER_LIBRARY NAMES libroctracer64.so HINTS
     ${ROCM_SOURCE_DIR}/lib)
-else()
-  find_library(ROCPROF_LIBRARY NAMES libroctracer64.so HINTS
-    ${ROCM_SOURCE_DIR}/lib)
-endif()
-  target_link_libraries(kineto "${ROCPROF_LIBRARY}")
+  target_link_libraries(kineto "${ROCTRACER_LIBRARY}")
   find_library(KINETO_HIP_LIBRARY NAMES libamdhip64.so HINTS
     ${ROCM_SOURCE_DIR}/lib)
   target_link_libraries(kineto "${KINETO_HIP_LIBRARY}")
