+++ libkineto/libkineto_defs.bzl
@@ -27,18 +27,10 @@ def get_libkineto_cupti_srcs(with_api = True):
         "src/cupti_strings.cpp",
     ] + (get_libkineto_cpu_only_srcs(with_api))
 
-def get_libkineto_rocprofiler_srcs(with_api = True):
-    return [
-        "src/RocprofActivityApi.cpp",
-        "src/RocprofLogger.cpp",
-        "src/RocLogger.cpp",
-    ] + (get_libkineto_cpu_only_srcs(with_api))
-
 def get_libkineto_roctracer_srcs(with_api = True):
     return [
         "src/RoctracerActivityApi.cpp",
         "src/RoctracerLogger.cpp",
-        "src/RocLogger.cpp",
     ] + (get_libkineto_cpu_only_srcs(with_api))
 
 def get_libkineto_xpupti_srcs(with_api = True):
