+++ libkineto/src/RoctracerActivityApi.cpp
@@ -40,7 +40,7 @@ void RoctracerActivityApi::pushCorrelationID(int id, CorrelationFlowType type) {
     return;
   }
   singleton().d->pushCorrelationID(
-      id, static_cast<RocLogger::CorrelationDomain>(type));
+      id, static_cast<RoctracerLogger::CorrelationDomain>(type));
 #endif
 }
 
@@ -50,7 +50,7 @@ void RoctracerActivityApi::popCorrelationID(CorrelationFlowType type) {
     return;
   }
   singleton().d->popCorrelationID(
-      static_cast<RocLogger::CorrelationDomain>(type));
+      static_cast<RoctracerLogger::CorrelationDomain>(type));
 #endif
 }
 
@@ -85,8 +85,8 @@ timestamp_t getTimeOffset() {
 }
 
 int RoctracerActivityApi::processActivities(
-    std::function<void(const rocprofBase*)> handler,
-    std::function<void(uint64_t, uint64_t, RocLogger::CorrelationDomain)>
+    std::function<void(const roctracerBase*)> handler,
+    std::function<void(uint64_t, uint64_t, RoctracerLogger::CorrelationDomain)>
         correlationHandler) {
   // Find offset to map from monotonic clock to system clock.
   // This will break time-ordering of events but is status quo.
@@ -94,15 +94,15 @@ int RoctracerActivityApi::processActivities(
   int count = 0;
 
   // Process all external correlations pairs
-  for (int it = RocLogger::CorrelationDomain::begin;
-       it < RocLogger::CorrelationDomain::end;
+  for (int it = RoctracerLogger::CorrelationDomain::begin;
+       it < RoctracerLogger::CorrelationDomain::end;
        ++it) {
     auto& externalCorrelations = d->externalCorrelations_[it];
     for (auto& item : externalCorrelations) {
       correlationHandler(
           item.first,
           item.second,
-          static_cast<RocLogger::CorrelationDomain>(it));
+          static_cast<RoctracerLogger::CorrelationDomain>(it));
     }
     std::lock_guard<std::mutex> lock(d->externalCorrelationsMutex_);
     externalCorrelations.clear();
@@ -122,7 +122,7 @@ int RoctracerActivityApi::processActivities(
         !isLogged(ActivityType::CUDA_RUNTIME)) {
       filtered = true;
     } else {
-      switch (reinterpret_cast<rocprofAsyncRow*>(item)->kind) {
+      switch (reinterpret_cast<roctracerAsyncRow*>(item)->kind) {
         case HIP_OP_COPY_KIND_DEVICE_TO_HOST_:
         case HIP_OP_COPY_KIND_HOST_TO_DEVICE_:
         case HIP_OP_COPY_KIND_DEVICE_TO_DEVICE_:
@@ -142,7 +142,8 @@ int RoctracerActivityApi::processActivities(
           if (!isLogged(ActivityType::CONCURRENT_KERNEL))
             filtered = true;
           // Don't record barriers/markers
-          if (reinterpret_cast<rocprofAsyncRow*>(item)->op == HIP_OP_ID_BARRIER)
+          if (reinterpret_cast<roctracerAsyncRow*>(item)->op ==
+              HIP_OP_ID_BARRIER)
             filtered = true;
           break;
       }
