+++ libkineto/src/RoctracerActivityApi.h
@@ -48,8 +48,9 @@ class RoctracerActivityApi {
   void setMaxEvents(uint32_t maxEvents);
 
   virtual int processActivities(
-      std::function<void(const rocprofBase*)> handler,
-      std::function<void(uint64_t, uint64_t, RocLogger::CorrelationDomain)>
+      std::function<void(const roctracerBase*)> handler,
+      std::function<
+          void(uint64_t, uint64_t, RoctracerLogger::CorrelationDomain)>
           correlationHandler);
 
   void setMaxBufferSize(int size);
