+++ libkineto/src/RoctracerLogger.h
@@ -24,8 +24,6 @@
 #include <roctracer_hip.h>
 #include <roctracer_roctx.h>
 
-#include "RocLogger.h"
-
 // Local copy of hip op types.  These are public (and stable) in later rocm
 // releases
 typedef enum {
@@ -58,13 +56,203 @@ namespace libkineto {
 class RoctracerActivityApi;
 }
 
-class RoctracerApiIdList : public ApiIdList {
+typedef uint64_t timestamp_t;
+
+static timestamp_t timespec_to_ns(const timespec& time) {
+  return ((timestamp_t)time.tv_sec * 1000000000) + time.tv_nsec;
+}
+
+class ApiIdList {
  public:
-  uint32_t mapName(const std::string& apiName) override;
+  ApiIdList();
+  bool invertMode() {
+    return invert_;
+  }
+  void setInvertMode(bool invert) {
+    invert_ = invert;
+  }
+  void add(const std::string& apiName);
+  void remove(const std::string& apiName);
+  bool loadUserPrefs();
+  bool contains(uint32_t apiId);
+  const std::unordered_map<uint32_t, uint32_t>& filterList() {
+    return filter_;
+  }
+
+ private:
+  std::unordered_map<uint32_t, uint32_t> filter_;
+  bool invert_;
+};
+
+typedef enum {
+  ROCTRACER_ACTIVITY_DEFAULT = 0,
+  ROCTRACER_ACTIVITY_KERNEL,
+  ROCTRACER_ACTIVITY_COPY,
+  ROCTRACER_ACTIVITY_MALLOC,
+  ROCTRACER_ACTIVITY_ASYNC,
+  ROCTRACER_ACTIVITY_NONE
+} roctracer_activity_types;
+
+struct roctracerBase {
+  roctracerBase(
+      uint64_t id,
+      uint32_t domain,
+      uint64_t begin,
+      uint64_t end,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_NONE)
+      : id(id), begin(begin), end(end), domain(domain), type(type) {}
+  uint64_t id; // correlation_id
+  uint64_t begin;
+  uint64_t end;
+  uint32_t domain;
+  roctracer_activity_types type;
+};
+
+struct roctracerRow : public roctracerBase {
+  roctracerRow(
+      uint64_t id,
+      uint32_t domain,
+      uint32_t cid,
+      uint32_t pid,
+      uint32_t tid,
+      uint64_t begin,
+      uint64_t end,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_DEFAULT)
+      : roctracerBase(id, domain, begin, end, type),
+        cid(cid),
+        pid(pid),
+        tid(tid) {}
+  uint32_t cid;
+  uint32_t pid;
+  uint32_t tid;
+};
+
+struct roctracerKernelRow : public roctracerRow {
+  roctracerKernelRow(
+      uint64_t id,
+      uint32_t domain,
+      uint32_t cid,
+      uint32_t pid,
+      uint32_t tid,
+      uint64_t begin,
+      uint64_t end,
+      const void* faddr,
+      hipFunction_t function,
+      unsigned int gx,
+      unsigned int gy,
+      unsigned int gz,
+      unsigned int wx,
+      unsigned int wy,
+      unsigned int wz,
+      size_t gss,
+      hipStream_t stream,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_KERNEL)
+      : roctracerRow(id, domain, cid, pid, tid, begin, end, type),
+        functionAddr(faddr),
+        function(function),
+        gridX(gx),
+        gridY(gy),
+        gridZ(gz),
+        workgroupX(wx),
+        workgroupY(wy),
+        workgroupZ(wz),
+        groupSegmentSize(gss),
+        stream(stream) {}
+  const void* functionAddr;
+  hipFunction_t function;
+  unsigned int gridX;
+  unsigned int gridY;
+  unsigned int gridZ;
+  unsigned int workgroupX;
+  unsigned int workgroupY;
+  unsigned int workgroupZ;
+  size_t groupSegmentSize;
+  hipStream_t stream;
+};
+
+struct roctracerCopyRow : public roctracerRow {
+  roctracerCopyRow(
+      uint64_t id,
+      uint32_t domain,
+      uint32_t cid,
+      uint32_t pid,
+      uint32_t tid,
+      uint64_t begin,
+      uint64_t end,
+      const void* src,
+      const void* dst,
+      size_t size,
+      hipMemcpyKind kind,
+      hipStream_t stream,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_COPY)
+      : roctracerRow(id, domain, cid, pid, tid, begin, end, type),
+        src(src),
+        dst(dst),
+        size(size),
+        kind(kind),
+        stream(stream) {}
+  const void* src;
+  const void* dst;
+  size_t size;
+  hipMemcpyKind kind;
+  hipStream_t stream;
+};
+
+struct roctracerMallocRow : public roctracerRow {
+  roctracerMallocRow(
+      uint64_t id,
+      uint32_t domain,
+      uint32_t cid,
+      uint32_t pid,
+      uint32_t tid,
+      uint64_t begin,
+      uint64_t end,
+      const void* ptr,
+      size_t size,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_MALLOC)
+      : roctracerRow(id, domain, cid, pid, tid, begin, end, type),
+        ptr(ptr),
+        size(size) {}
+  const void* ptr;
+  size_t size;
+};
+
+struct roctracerAsyncRow : public roctracerBase {
+  roctracerAsyncRow(
+      uint64_t id,
+      uint32_t domain,
+      uint32_t kind,
+      uint32_t op,
+      int device,
+      uint64_t queue,
+      uint64_t begin,
+      uint64_t end,
+      const std::string& kernelName,
+      roctracer_activity_types type = ROCTRACER_ACTIVITY_ASYNC)
+      : roctracerBase(id, domain, begin, end, type),
+        kind(kind),
+        op(op),
+        device(device),
+        queue(queue),
+        kernelName(kernelName) {}
+  uint32_t kind;
+  uint32_t op;
+  int device;
+  uint64_t queue;
+  std::string kernelName;
 };
 
 class RoctracerLogger {
  public:
+  enum CorrelationDomain {
+    begin,
+    Default = begin,
+    Domain0 = begin,
+    Domain1,
+    end,
+    size = end
+  };
+
   RoctracerLogger();
   RoctracerLogger(const RoctracerLogger&) = delete;
   RoctracerLogger& operator=(const RoctracerLogger&) = delete;
@@ -73,8 +261,8 @@ class RoctracerLogger {
 
   static RoctracerLogger& singleton();
 
-  static void pushCorrelationID(uint64_t id, RocLogger::CorrelationDomain type);
-  static void popCorrelationID(RocLogger::CorrelationDomain type);
+  static void pushCorrelationID(uint64_t id, CorrelationDomain type);
+  static void popCorrelationID(CorrelationDomain type);
 
   void startLogging();
   void stopLogging();
@@ -86,7 +274,7 @@ class RoctracerLogger {
   void endTracing();
 
   roctracer_pool_t* hccPool_{NULL};
-  static void insert_row_to_buffer(rocprofBase* row);
+  static void insert_row_to_buffer(roctracerBase* row);
   static void api_callback(
       uint32_t domain,
       uint32_t cid,
@@ -94,18 +282,18 @@ class RoctracerLogger {
       void* arg);
   static void activity_callback(const char* begin, const char* end, void* arg);
 
-  RoctracerApiIdList loggedIds_;
+  ApiIdList loggedIds_;
 
   // Api callback data
   uint32_t maxBufferSize_{5000000}; // 5M GPU runtime/kernel events.
-  std::vector<rocprofBase*> rows_;
+  std::vector<roctracerBase*> rows_;
   std::mutex rowsMutex_;
 
   // This vector collects pairs of correlationId and their respective
   // externalCorrelationId for each CorrelationDomain. This will be used
   // to populate the Correlation maps during post processing.
   std::vector<std::pair<uint64_t, uint64_t>>
-      externalCorrelations_[RocLogger::CorrelationDomain::size];
+      externalCorrelations_[CorrelationDomain::size];
   std::mutex externalCorrelationsMutex_;
 
   bool externalCorrelationEnabled_{true};
